Admin Router: Refactor lua files into the new layout
