+++ packages/adminrouter/extra/src/common/http.conf
@@ -5,10 +5,10 @@ default_type application/octet-stream;
 sendfile on;
 keepalive_timeout 65;
 
-lua_package_path '$prefix/conf/?.lua;;';
+lua_package_path '$prefix/conf/lib/?.lua;;';
 
 init_by_lua_block {
-    util = require "common.util"
+    util = require "util"
 
     DEFAULT_SCHEME = os.getenv("DEFAULT_SCHEME")
     if DEFAULT_SCHEME == nil then
@@ -39,17 +39,17 @@ init_by_lua_block {
     end
     ngx.log(ngx.NOTICE, "Marathon upstream: " .. UPSTREAM_MARATHON)
 
+    -- Initialize modules:
+    cache = require "cache".init(nil);
+    service = require "service".init();
+    mleader = require "marathon-leader".init();
+    agent = require "agent".init();
+    historyservice = require "history-service".init();
     -- Loading the auth module in the global Lua VM in the master process is a
     -- requirement, so that reading the secret is executed under the user that
     -- spawns the master process instead of 'nobody' (which workers operate
     -- under).
-    auth = require "common.auth.open".init(os.getenv("ADMINROUTER_ACTIVATE_AUTH_MODULE"));
-    -- Initialize other modules:
-    cache = require "master.cache".init(nil);
-    service = require "master.service".init();
-    mleader = require "master.marathon-leader".init();
-    agent = require "master.agent".init();
-    historyservice = require "master.history-service".init();
+    auth = require "auth.open".init(os.getenv("ADMINROUTER_ACTIVATE_AUTH_MODULE"));
 }
 
 # Name: DC/OS Component Package Manager (Pkgpanda)
