+++ packages/adminrouter/extra/src/nginx.master.conf
@@ -285,7 +285,7 @@ http {
             access_by_lua_block {
                 auth.access_metadata_endpoint();
             }
-            content_by_lua_file conf/master/metadata.lua;
+            content_by_lua_file conf/lib/metadata.lua;
         }
 
         # TODO: split this into its own file
